blobAvScan/scanBlob/requirements.txt
asn1crypto==0.24.0
azure-common==1.1.14
azure-functions==1.0.0a4
azure-functions-worker==1.0.0a4
azure-nspkg==2.0.0
azure-storage-blob==1.3.1
azure-storage-common==1.3.0
azure-storage-nspkg==3.0.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
grpcio==1.14.2
grpcio-tools==1.14.2
idna==2.7
protobuf==3.6.1
pyClamd==0.4.0
pycparser==2.18
python-dateutil==2.7.3
requests==2.19.1
six==1.11.0
urllib3==1.23
